improve implementation of the update
