+++ core/src/main/scala/org/apache/spark/sql/rapids/tool/store/AccMetaRef.scala
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2024, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql.rapids.tool.store
+
+case class AccMetaRef(id: Long, name: AccNameRef) {
+
+}
+
+object AccMetaRef {
+  def apply(id: Long, name: Option[String]): AccMetaRef =
+    new AccMetaRef(id, AccNameRef.internAccName(name))
+}
